split localities map (not implemented) and coordinate selection map
