+++ frontend/src/components/Map/CoordinateSelectionMap.tsx
