+++ frontend/src/components/MapPage.tsx
@@ -1,9 +1,6 @@
-import { Map } from './Map/Map'
-import { useState } from 'react'
+import { LocalitiesMap } from './Map/LocalitiesMap'
 import 'leaflet/dist/leaflet.css'
 
 export const MapPage = () => {
-  const [coordinates, setCoordinates] = useState({ lat: 60.202665856, lng: 24.957662836 })
-
-  return <Map coordinates={coordinates} setCoordinates={setCoordinates} />
+  return <LocalitiesMap />
 }
